Simplify `bevy_utils` features
