+++ crates/bevy_state/Cargo.toml
@@ -30,7 +30,6 @@ bevy_app = ["dep:bevy_app"]
 ## supported platforms.
 std = [
   "bevy_ecs/std",
-  "bevy_utils/std",
   "bevy_reflect?/std",
   "bevy_app?/std",
   "bevy_platform/std",
@@ -40,7 +39,6 @@ std = [
 ## on all platforms, including `no_std`.
 critical-section = [
   "bevy_ecs/critical-section",
-  "bevy_utils/critical-section",
   "bevy_app?/critical-section",
   "bevy_reflect?/critical-section",
   "bevy_platform/critical-section",
